brush barebones added

+++ renderer/square.cpp
@@ -16,8 +16,7 @@
 #endif
 #include <iostream>
 
-class Canvas
-{
+class Canvas {
 public:
     int w;
     int h;
@@ -27,58 +26,55 @@ public:
     int y2;
     bool drawing;
     std::string tool = "brush";
-    SDL_Surface* surface; 
-    std::array<int, 3> bodycolor;
+    SDL_Surface* surface;
+    std::array<float, 3> bodycolor;
 
     Canvas(int x1 = 0,
-           int y1 = 0,
-           int x2 = 0,
-           int y2 = 0,
-           bool drawing = false,
-           std::array<int, 3> bodycolor = {255, 255, 255})
-        : x1(x1), y1(y1), x2(x2), y2(y2), drawing(drawing), bodycolor(bodycolor), surface(SDL_CreateRGBSurface(0, w, h, 24, 0, 0, 0, 0)) {}
+        int y1 = 0,
+        int x2 = 0,
+        int y2 = 0,
+        bool drawing = false,
+        std::array<float, 3> bodycolor = { 255, 255, 255 })
+        : x1(x1), y1(y1), x2(x2), y2(y2), drawing(drawing), bodycolor(bodycolor), surface(SDL_CreateRGBSurface(0, w, h, 24, 0, 0, 0, 0)) {
+    }
 
     void draw();
 };
 
 
-void Canvas::draw()
-{
+void Canvas::draw() {
     std::cout << "It works" << std::endl;
 }
 
 /*public:
-    Canvas(int w, int h, std::array<int, 3> &bodycolor) : w(w), h(h), bodycolor(bodycolor)
+    Canvas(int w, int h, std::array<float, 3> &bodycolor) : w(w), h(h), bodycolor(bodycolor)
     {
 
-     /*    std::array<int, 3> getColor() const
+     /*    std::array<float, 3> getColor() const
         {
             return bodycolor;
         }
     }
 };
 */
-static void DrawCanvas(SDL_Renderer *renderer, int x, int y)
-{
+static void DrawCanvas(SDL_Renderer* renderer, int x, int y) {
     // Draw Shape
-    SDL_Rect Canvas = {x, y, 200, 200};
+    SDL_Rect Canvas = { x, y, 200, 200 };
     SDL_SetRenderDrawColor(renderer, 255, 0, 0, 255);
     SDL_RenderFillRect(renderer, &Canvas);
     SDL_RenderPresent(renderer);
 }
 
 // Main code
-int main(int, char **)
-{
+int main(int, char**) {
     // Setup SDL
-    if (SDL_Init(SDL_INIT_VIDEO | SDL_INIT_TIMER | SDL_INIT_GAMECONTROLLER) != 0)
-    {
+    if (SDL_Init(SDL_INIT_VIDEO | SDL_INIT_TIMER | SDL_INIT_GAMECONTROLLER) != 0) {
         printf("Error: %s\n", SDL_GetError());
         return -1;
     }
 
     // GL 3.0 + GLSL 130
-    const char *glsl_version = "#version 130";
+    const char* glsl_version = "#version 130";
     SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, 0);
     SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
     SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
@@ -89,11 +85,10 @@ int main(int, char **)
     SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24);
     SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, 8);
     SDL_WindowFlags window_flags = (SDL_WindowFlags)(SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE | SDL_WINDOW_ALLOW_HIGHDPI);
-    SDL_Window *window = SDL_CreateWindow("Sumi", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, 1280, 720, window_flags);
-    SDL_Renderer *renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
+    SDL_Window* window = SDL_CreateWindow("Sumi", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, 1280, 720, window_flags);
+    SDL_Renderer* renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
 
-    if (window == nullptr)
-    {
+    if (window == nullptr) {
         printf("Error: SDL_CreateWindow(): %s\n", SDL_GetError());
         return -1;
     }
@@ -103,14 +98,14 @@ int main(int, char **)
     SDL_GL_SetSwapInterval(1); // Enable vsync
 
     // Set window icon
-    SDL_Surface *iconSurface = IMG_Load("../icons/icon.png");
+    SDL_Surface* iconSurface = IMG_Load("../icons/icon.png");
     SDL_SetWindowIcon(window, iconSurface);
     SDL_FreeSurface(iconSurface);
 
     // Setup Dear ImGui context
     IMGUI_CHECKVERSION();
     ImGui::CreateContext();
-    ImGuiIO &io = ImGui::GetIO();
+    ImGuiIO& io = ImGui::GetIO();
     (void)io;
     io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
     io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;  // Enable Gamepad Controls
@@ -127,22 +122,18 @@ int main(int, char **)
 
     bool done = false;
 
-    while (!done)
-    {
+    while (!done) {
         SDL_Event event;
 
-        while (SDL_PollEvent(&event))
-        {
+        while (SDL_PollEvent(&event)) {
             ImGui_ImplSDL2_ProcessEvent(&event);
 
-            switch (event.type)
-            { // Handles user inputs
+            switch (event.type) { // Handles user inputs
             case SDL_QUIT:
                 done = true;
 
             case SDL_MOUSEBUTTONDOWN:
-                switch (event.button.button)
-                {
+                switch (event.button.button) {
                 case SDL_BUTTON_LEFT:
                     //  x1 = event.motion.x;
                     //  y1 = event.motion.y;
@@ -155,8 +146,7 @@ int main(int, char **)
                 break;
 
             case SDL_MOUSEBUTTONUP:
-                switch (event.button.button)
-                {
+                switch (event.button.button) {
                 case SDL_BUTTON_LEFT:
                     //     drawing = false;
                     break;
@@ -181,16 +171,15 @@ int main(int, char **)
         // Rendering
         ImGui::Render();
         glViewport(0, 0, (int)io.DisplaySize.x, (int)io.DisplaySize.y);
-        ImGuiStyle &style = ImGui::GetStyle();
+        ImGuiStyle& style = ImGui::GetStyle();
         ImVec4 clearColor = ImGui::ColorConvertU32ToFloat4(Default::GRAY100);
         glClearColor(clearColor.x, clearColor.y, clearColor.z, clearColor.w);
         glClear(GL_COLOR_BUFFER_BIT);
         ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
 
         // Update and Render additional Platform Windows
-        if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
-        {
-            SDL_Window *backup_current_window = SDL_GL_GetCurrentWindow();
+        if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable) {
+            SDL_Window* backup_current_window = SDL_GL_GetCurrentWindow();
             SDL_GLContext backup_current_context = SDL_GL_GetCurrentContext();
             ImGui::UpdatePlatformWindows();
             ImGui::RenderPlatformWindowsDefault();
